feat: initial kubectl layer construct (#1)
This is nearly a direct copy from https://github.com/aws/aws-cdk/tree/main/packages/%40aws-cdk/lambda-layer-kubectl with some changes to make it v2-compatible and projen-compatible.

Note: this exposes a `kubeCtlLayer` construct, not the final form `kubeCtlAsset`. I figure it is better to start here and then update this repo in subsequent PRs. However, that means there are some inconsistencies with what I say this library exposes -- since this is not meant for active use right now, I feel like that is ok and will be corrected in a future PR.