feat: expose `LAYER_SOURCE` and `ASSET_PATH` consts (#52)
Removes the asset in v20 in favor of two constants: `LAYER_SOURCE_DIR` and `ASSET_PATH`. These constants are meant to be consumed by aws-cdk-lib. Since it is to be consumed by aws-cdk-lib, it cannot depend on aws-cdk-lib, as a consequence, aws-cdk-lib is no longer a peer dependency and instead a dev dependency for testing.