fix: bump deps to make PyPI distribution PEP 625 compliant (#1635)
With the dependencies in the previous release, the PyPI upload was named `aws-cdk.lambda-layer-kubectl-v30-2.0.2.tar.gz`, when it should have been named `aws_cdk_lambda_layer_kubectl_v30-2.0.2.tar.gz`. By bumping the dependencies we use for releasing and triggering a release, the next version should be compliant with this normalized name.