chore: script to automate new Kubectl version releases (#1731)
Today, we have to follow a list of manual steps when bumping to the latest version of Kubectl. However, since many of the steps can be automated, I wrote this script to simplify the process and make it (nearly) fool-proof.

See [this commit](https://github.com/Leo10Gama/awscdk-asset-kubectl/commit/dbe12880fac1ab3f5064e4dac1368cf76e7cc891) for a sample result of the script in action. (or run it locally!)